refactor: add support for env variable for 3d party API address
